moved all classes in dedicated SPI packages in identity-hub-spi
