+++ spi/identity-hub-spi/src/main/java/org/eclipse/edc/identityhub/spi/authorization/AuthorizationResultHandler.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2024 Metaform Systems, Inc.
+ *  Copyright (c) 2025 Cofinity-X
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,11 +8,10 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Metaform Systems, Inc. - initial API and implementation
- *
+ *       Cofinity-X - initial API and implementation
  */
 
-package org.eclipse.edc.identityhub.spi;
+package org.eclipse.edc.identityhub.spi.authorization;
 
 import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.result.ServiceFailure;
