+++ spi/identity-hub-spi/src/main/java/org/eclipse/edc/identityhub/spi/transformation/ScopeToCriterionTransformer.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *  Copyright (c) 2025 Cofinity-X
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,11 +8,10 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
+ *       Cofinity-X - initial API and implementation
  */
 
-package org.eclipse.edc.identityhub.spi;
+package org.eclipse.edc.identityhub.spi.transformation;
 
 import org.eclipse.edc.spi.query.Criterion;
 import org.eclipse.edc.spi.result.Result;
